# gateway/gateway-controller/api/management-openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Controller Management API
  description: |
    REST API for managing API configurations in the WSO2 API Platform Gateway.
  version: 0.9.0
  contact:
    name: WSO2 API Platform Team
servers:
  - url: http://localhost:9090/api/management/v0.9
    description: Local development
  - url: http://gateway-controller:9090/api/management/v0.9
    description: Docker/Kubernetes deployment
security:
  - basicAuth: []
paths:
...
  /rest-apis/{id}/api-keys:
    post:
      summary: Create a new API key for an API
      description: Generate a new API key for a RestAPI in the Gateway. The key is a 32-byte random value encoded in hexadecimal, prefixed with `apip_`. Use the API Key policy on the API to validate incoming requests with this key.
      operationId: createAPIKey
      x-basicauth-roles: [admin, consumer]
      tags:
        - Rest API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to generate the key for
          schema:
            type: string
          example: reading-list-api-v1.0
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/APIKeyCreationRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/APIKeyCreationRequest"
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreationResponse"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: RestAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: Get the list of API keys for an API
      description: List all API keys for a RestAPI in the Gateway.
      operationId: listAPIKeys
      x-basicauth-roles: [admin, consumer]
      tags:
        - Rest API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to retrieve the keys for
          schema:
            type: string
          example: reading-list-api-v1.0
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyListResponse"
        "404":
          description: RestAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rest-apis/{id}/api-keys/{apiKeyName}/regenerate:
    post:
      summary: Regenerate API key for an API
      description: Regenerate an existing API key for a RestAPI in the Gateway. The previous key is revoked and replaced with a new 32-byte random value encoded in hexadecimal, prefixed with `apip_`.
      operationId: regenerateAPIKey
      x-basicauth-roles: [admin, consumer]
      tags:
        - Rest API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to generate the key for
          schema:
            type: string
          example: reading-list-api-v1.0
        - name: apiKeyName
          in: path
          required: true
          description: |
            Name of the API key to regenerate
          schema:
            type: string
          example: reading-list-api-key
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/APIKeyRegenerationRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/APIKeyRegenerationRequest"
      responses:
        '200':
          description: API key rotated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreationResponse"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: RestAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rest-apis/{id}/api-keys/{apiKeyName}:
    put:
      summary: Update an API key with a new regenerated value
      description: Update an API key with a custom value instead of auto-generating one.
      operationId: updateAPIKey
      x-basicauth-roles: [admin, consumer]
      tags:
        - Rest API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API
          schema:
            type: string
          example: reading-list-api-v1.0
        - name: apiKeyName
          in: path
          required: true
          description: |
            Name of the API key to update
          schema:
            type: string
          example: reading-list-api-key
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/APIKeyUpdateRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/APIKeyUpdateRequest"
      responses:
        '200':
          description: API key updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreationResponse"
        "400":
          description: Invalid request (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: API or API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Revoke an API key
      description: Revoke an API key. Once revoked, it can no longer be used to authenticate requests.
      operationId: revokeAPIKey
      x-basicauth-roles: [admin, consumer]
      tags:
        - Rest API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to revoke the key for
          schema:
            type: string
          example: reading-list-api-v1.0
        - name: apiKeyName
          in: path
          required: true
          description: |
            Name of the API key to revoke
          schema:
            type: string
          example: reading-list-api-key
      responses:
        '200':
          description: API key revoked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyRevocationResponse"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: RestAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # SOAP API Management
  # ---------------------------------------------------------------------------
  /soap-apis:
    post:
      summary: Create a new SoapAPI
      description: Add a new SOAP API to the Gateway.
      operationId: createSoapAPI
      x-basicauth-roles: [admin, developer]
      tags:
        - SOAP API Management
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/SoapAPIRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/SoapAPIRequest"
      responses:
        "201":
          description: SoapAPI created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoapAPI"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - API with same name and version already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: List all SoapAPIs
      description: List SOAP APIs registered in the Gateway, optionally filtered by name, version, context, or status.
      operationId: listSoapAPIs
      x-basicauth-roles: [admin, developer]
      tags:
        - SOAP API Management
      parameters:
        - name: displayName
          in: query
          required: false
          description: Filter by API display name
          schema:
            type: string
        - name: version
          in: query
          required: false
          description: Filter by API version
          schema:
            type: string
        - name: context
          in: query
          required: false
          description: Filter by API context path
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Filter by deployment status
          schema:
            type: string
            enum: [deployed, undeployed]
      responses:
        "200":
          description: List of SoapAPIs
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  count:
                    type: integer
                  soapApis:
                    type: array
                    items:
                      $ref: "#/components/schemas/SoapAPI"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /soap-apis/{id}:
    get:
      summary: Get SoapAPI by id
      description: Get a SOAP API by its ID.
      operationId: getSoapAPIById
      x-basicauth-roles: [admin, developer]
      tags:
        - SOAP API Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SoapAPI details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoapAPI"
            application/yaml:
              schema:
                $ref: "#/components/schemas/SoapAPI"
        "404":
          description: SoapAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Update an existing SoapAPI
      description: Update an existing SOAP API in the Gateway.
      operationId: updateSoapAPI
      x-basicauth-roles: [admin, developer]
      tags:
        - SOAP API Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/SoapAPIRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/SoapAPIRequest"
      responses:
        "200":
          description: SoapAPI updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoapAPI"
        "400":
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: SoapAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete a SoapAPI
      description: Delete a SOAP API from the Gateway.
      operationId: deleteSoapAPI
      x-basicauth-roles: [admin, developer]
      tags:
        - SOAP API Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SoapAPI deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: SoapAPI deleted successfully
                  id:
                    type: string
        "404":
          description: SoapAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /soap-apis/{id}/api-keys:
    post:
      summary: Generate a new API key for a SoapAPI
      description: Generate a new API key for a SOAP API in the Gateway.
      operationId: createSoapAPIKey
      x-basicauth-roles: [admin, developer]
      tags:
        - SOAP API Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/APIKeyCreationRequest"
      responses:
        "201":
          description: API key created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreationResponse"
        "404":
          description: SoapAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: List all API keys for a SoapAPI
      description: List all API keys for a SOAP API in the Gateway.
      operationId: listSoapAPIKeys
      x-basicauth-roles: [admin, developer]
      tags:
        - SOAP API Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyListResponse"
        "404":
          description: SoapAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /soap-apis/{id}/api-keys/{apiKeyName}/regenerate:
    post:
      summary: Regenerate an existing API key for a SoapAPI
      description: Regenerate an existing API key for a SOAP API.
      operationId: regenerateSoapAPIKey
      x-basicauth-roles: [admin, developer]
      tags:
        - SOAP API Management
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: apiKeyName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/APIKeyRegenerationRequest"
      responses:
        "200":
          description: API key regenerated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreationResponse"
        "404":
          description: SoapAPI or API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    APIKeyCreationRequest:
      type: object
      properties:
        name:
          type: string
          description: Identifier of the API key. If not provided, a default identifier will be generated
          pattern: "^[a-z0-9]+(-[a-z0-9]+)*$"
          minLength: 3
          maxLength: 63
          example: my-production-key
        apiKey:
          type: string
          minLength: 36
          description: |
            Optional plain-text API key value for external key injection.
            If provided, this key will be used instead of generating a new one.
            The key will be hashed before storage. The key can be in any format
            (minimum 36 characters). Use this for injecting externally generated
            API keys.
          example: "xxxxxx-wso2-api-platform-key-xxxxxx-xxxxxxx"
        maskedApiKey:
          type: string
          description: |
            Masked version of the API key for display purposes.
            Provided by the platform API when injecting pre-hashed keys.
          example: "apip_****xyz789"
        expiresIn:
          type: object
          description: Expiration duration for the API key
          properties:
            unit:
              type: string
              description: Time unit for expiration
              enum:
                - seconds
                - minutes
                - hours
                - days
                - weeks
                - months
              example: days
            duration:
              type: integer
              description: Duration value for expiration
              example: 30
          required:
            - unit
            - duration
        expiresAt:
          type: string
          format: date-time
          description: Expiration timestamp. If both expiresIn and expiresAt are provided, expiresAt takes precedence.
          example: "2026-12-08T10:30:00Z"
        externalRefId:
          type: string
          description: |
            External reference ID for the API key.
            This field is optional and used for tracing purposes only.
            The gateway generates its own internal ID for tracking.
          example: "cloud-apim-key-98765"
        issuer:
          type: string
          description: |
            Identifies the portal that created this key. If provided, only api keys generated from
            the same portal will be accepted. If not provided, there is no portal restriction.
          example: "api-platform-devportal"
      example:
        name: my-production-key
...